Refactor: use MUI components for Quest and addBidModal comoonents
